upgrades: Move auto upgrades configure form to first tab. Split forms and views into separate files.
