+++ plinth/modules/upgrades/templates/upgrades_configure.html
@@ -22,6 +22,8 @@
 
 {% block content %}
 
+  <h2>{{ title }}</h2>
+
   <form class="form" method="post">
     {% csrf_token %}
